Complete major codebase reorganization: move infrastructure to core/ module
## 🏗️ **Major Structural Improvements**

### ✅ **Files Moved to Core Module**
- `config.rs` → `core/config.rs` (Application configuration)
- `database.rs` → `core/database.rs` (Database connections & migrations)
- `error.rs` → `core/error.rs` (Error handling & HTTP responses)
- `server.rs` → `core/server.rs` (HTTP server & routing setup)
- `openapi.rs` → `core/openapi.rs` (API documentation generation)

### ✅ **Files Removed for Clean Architecture**
- ❌ `core/models.rs` (removed - only contained re-exports)
- ❌ `types.rs` (removed - re-exports create import confusion)

### ✅ **Better Domain Organization**
- Auth models (Session, ApiKey) moved from core to `auth/models.rs`
- Clean module boundaries with proper separation of concerns
- No more mixed infrastructure and domain logic

### 🎯 **Smart Import Strategy**
- **lib.rs exports**: Common types (`Error`, `AppState`, `Result`) available as `crate::Error`
- **Direct imports**: Use proper module paths instead of confusing re-exports
- **Clean boundaries**: Core infrastructure separate from domain logic

### 🏗️ **New Clean Structure**
```
starter/src/
├── core/              # Infrastructure & shared utilities
│   ├── config.rs      # App configuration
│   ├── database.rs    # Database connections & migrations
│   ├── error.rs       # Error handling & HTTP responses
│   ├── openapi.rs     # API documentation generation
│   ├── server.rs      # HTTP server & routing setup
│   ├── state.rs       # Application state management
│   └── types.rs       # Core type definitions
├── auth/              # Authentication, sessions & API keys
├── users/             # User management domain
├── tasks/             # Background task processing
├── monitoring/        # Observability & metrics
├── rbac/              # Role-based access control
├── health/            # Health checks & probes
└── api/               # API response & pagination types
```

### ⚡ **Benefits Achieved**
- **Cleaner imports**: `use crate::Error` instead of `use crate::core::error::Error`
- **Logical organization**: Infrastructure in `core/`, domain logic in feature modules
- **Easier navigation**: Related functionality properly grouped
- **Better maintainability**: Clear module ownership and responsibilities
- **Follows Rust best practices**: Proper separation of concerns

All tests updated and passing. Frontend integration maintains full functionality.
