+++ starter/src/auth/mod.rs
@@ -5,3 +5,4 @@ pub mod models;
 pub mod services;
 
 pub use middleware::{AuthUser, admin_middleware, auth_middleware, security_headers_middleware};
+pub use models::{ApiKey, Session, LoginRequest, RegisterRequest, LoginResponse, RefreshResponse};
