+++ starter/src/lib.rs
@@ -1,22 +1,18 @@
 pub mod api;
 pub mod auth;
 pub mod cli;
-pub mod config;
 pub mod core;
-pub mod database;
-pub mod error;
 pub mod health;
-pub mod models;
 pub mod monitoring;
-pub mod openapi;
 pub mod rbac;
-pub mod server;
 pub mod tasks;
-pub mod types;
 pub mod users;
 
-// Re-export common types
-pub use config::AppConfig;
-pub use database::Database;
-pub use error::Error;
-pub use types::{DbConn, DbPool, Result};
+// Re-export most commonly used core types for convenience
+pub use core::{
+    config::AppConfig,
+    database::Database,
+    error::Error,
+    state::AppState,
+    types::{DbConn, DbPool, Result},
+};
