+++ starter/src/rbac/middleware.rs
@@ -1,5 +1,5 @@
 use crate::auth::AuthUser;
-use crate::error::Error;
+use crate::Error;
 use crate::rbac::models::{Permission, Resource, UserRole};
 use crate::rbac::services;
 use axum::{extract::Request, middleware::Next, response::Response};
