+++ starter/src/tasks/api.rs
@@ -7,14 +7,14 @@ use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
 use crate::{
+    api::{ApiResponse, ErrorResponse},
     auth::AuthUser,
-    error::Error,
+    AppState, Error,
     rbac::services as rbac_services,
     tasks::{
         processor::TaskProcessor,
         types::{CreateTaskRequest, TaskFilter, TaskPriority, TaskResponse, TaskStats, TaskStatus},
     },
-    types::{ApiResponse, AppState, ErrorResponse},
 };
 
 #[derive(Debug, Serialize, Deserialize, utoipa::ToSchema)]
