+++ starter/src/users/services.rs
@@ -1,8 +1,7 @@
 use crate::rbac::UserRole;
 use crate::users::models::{CreateUserRequest, User, UserProfile};
 use crate::{
-    error::Error,
-    types::{DbConn, Result},
+    DbConn, Error, Result,
 };
 use argon2::password_hash::{SaltString, rand_core::OsRng};
 use argon2::{Argon2, PasswordHash, PasswordHasher, PasswordVerifier};
